+++ wise-office-client/src/components/project/addLog.ts
@@ -1,0 +1,5 @@
+const addLog = () => {
+    console.log("로그 작성");
+};
+
+export { addLog };
