+++ wise-office-client/src/components/project/index.ts
@@ -1,0 +1,6 @@
+export { addLog } from "./addLog";
+
+export { default as ProjectParticipants } from "./projectParticipants";
+export { default as ProjectParticipantItem } from "./projectParticipantItem";
+
+export * from "./types";
