+++ wise-office-client/src/components/project/types.ts
@@ -1,0 +1,15 @@
+export interface LogEntry {
+    id: number;
+    title: string;
+    user: string;
+    content: string;
+    date: string;
+    comments: Comment[];
+}
+
+export interface Comment {
+    id: number;
+    user: string;
+    content: string;
+    date: string;
+}
